AI/requirements.txt
transformers
python-dotenv
torch
peft
datasets
Flask
pyttsx3
speechrecognition
langdetect